Function parallel::divide

pub fn divide<T, F>(data: &mut [T], granularity: usize, operation: F) where T: Send, F: Fn(&mut [T], usize) + Sync

Parallelizes an operation over a mutable slice

The data will be divided in chunks of granularity size. A new thread will be spawned to "operate" over each chunk.

operation will receive two arguments:

The mutable chunk of data, and

The offset of this chunk from the start of data

Panics

Panics if any of the underlying threads panics

Example

Parallel map

extern crate parallel;
extern crate rand;

use rand::{Rng, XorShiftRng};
use std::num::Float;

let ref mut rng: XorShiftRng = rand::thread_rng().gen();
let mut v = (0..1_000).map(|_| rng.gen::<f32>()).collect::<Vec<_>>();
parallel::divide(v.as_mut_slice(), 100, |data, _| {
    for x in data.iter_mut() {
        *x = x.sin();
    }
});